requirements.txt
langchain>=1.0
langchain-huggingface
langchain-text-splitters
langchain-community
faiss-cpu
Markdown
sentence-transformers
lazy_loader
rank_bm25
jieba
openai
fastapi
uvicorn[standard]
python-dotenv